requirements.txt
asgiref==3.5.2
bleach==5.0.1
cart==1.2.2
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
cryptography==39.0.2
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
dj-stripe==2.8.1
Django==4.1.1
django-admin-filter==0.7
django-admin-report==0.6.2
django-ckeditor==6.7.0
django-crispy-forms==1.14.0
django-daraja==1.2.0
django-environ==0.10.0
django-filter==22.1
django-import-export==2.8.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-reports-admin==1.1.0
django-s3-sqlite==0.0.5
django-shopping-cart==0.1
django-summernote==0.8.20.0
et-xmlfile==1.1.0
filelock==3.8.0
gunicorn==20.1.0
idna==3.4
Jinja2==3.1.2
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==2.1.1
martor==1.6.26
odfpy==1.4.1
openpyxl==3.0.10
paymentsds-mpesa==0.1a10
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.6
pycparser==2.21
pycryptodome==3.17
python-decouple==3.8
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.2
stripe==4.2.0
tablib==3.2.1
typing_extensions==4.7.1
tzdata==2022.2
urllib3==1.22
virtualenv==20.16.5
webencodings==0.5.1
whitenoise==6.5.0
xlrd==2.0.1
xlwt==1.3.0
